Adds MovingAvarageArray version, and moves the older one to the package.
