+++ src/main/java/algorithms/curated170/easy/movingavarage/MovingAverage.java
@@ -1,4 +1,4 @@
-package algorithms.curated170.easy;
+package algorithms.curated170.easy.movingavarage;
 
 public class MovingAverage {
